+++ gems/aws-sigv4/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Support signature V4 signing per event
+
 1.0.3 (2018-06-28)
 ------------------